Move reusable functions to utils
