+++ python/turbine_models/custom_models/sd_inference/clip_test.py
@@ -13,11 +13,11 @@ import iree.compiler as ireec
 from iree.compiler.ir import Context
 import numpy as np
 from shark_turbine.aot import *
+from utils import *
 import torch
 import torch._dynamo as dynamo
 from transformers import CLIPTextModel, CLIPTokenizer
 
-import safetensors
 import argparse
 
 parser = argparse.ArgumentParser()
@@ -42,22 +42,6 @@ parser.add_argument(
 )
 
 prompt = ["a photograph of an astronaut riding a horse"]
-
-
-def save_external_weights(
-    mapper,
-    model,
-    external_weights=None,
-    external_weight_file=None,
-):
-    if external_weights is not None:
-        if external_weights == "safetensors":
-            mod_params = dict(model.named_parameters())
-            for name in mod_params:
-                mapper["params." + name] = name
-            if external_weight_file:
-                safetensors.torch.save_file(mod_params, external_weight_file)
-                print("Saved params to", external_weight_file)
 
 
 def export_clip_model(args):
@@ -129,12 +113,6 @@ def export_clip_model(args):
             f.write(flatbuffer_blob)
         print("Saved to", safe_name + ".vmfb")
         exit()
-
-
-def largest_error(array1, array2):
-    absolute_diff = np.abs(array1 - array2)
-    max_error = np.max(absolute_diff)
-    return max_error
 
 
 def run_clip_vmfb_comparison(args):
